command_java/Tests/tests/menu/parser/ResizeImageTest.java
package menu.parser;

import menu.action.IMenuCommandsFactory;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.ValueSource;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;

import static org.junit.jupiter.api.Assertions.*;
import static org.mockito.ArgumentMatchers.argThat;
import static org.mockito.ArgumentMatchers.eq;
import static org.mockito.Mockito.verify;
import static org.mockito.Mockito.verifyNoMoreInteractions;
import static org.mockito.Mockito.verifyZeroInteractions;
import static testUtils.Utils.assertNoThrows;

class ResizeImageTest {

    private ResizeImage _parser;
    @Mock
    private IMenuCommandsFactory _factory;

    @BeforeEach
    void setUp() {
        MockitoAnnotations.initMocks(this);
        _parser = new ResizeImage(_factory);
    }

    @ParameterizedTest
    @ValueSource(strings = { "", "test", "a", "1a", " 2 ", "1 ", "ResizeImage a", "ResizeImage 1 w", "ResizeImage 1 h", " ResizeImage 1 2 3" })
    void invalidCommandNoThrow(String text) {
        assertNoThrows(() -> assertNull(_parser.Parse(text)));
        verifyZeroInteractions(_factory);
    }

    @ParameterizedTest
    @ValueSource(strings = { "ResizeImage 1 2 3" })
    void validCommandResize(String text) {
        assertNoThrows(() -> _parser.Parse(text));
        verify(_factory).ResizeImage(
                eq(1),
                argThat((sz) -> sz.Height() == 3 && sz.Width() == 2));
        verifyNoMoreInteractions(_factory);
    }
}